Initial setup (#8)
Questions:
1. Can we discuss `./tests/testhat/setup.R`? I am not sure I understand
what it's doing.
    - I can delete this file.
3. What is the reporter here? `test_check(pkg_name, reporter =
ParallelProgressReporter$new())`
- Is this similar to the parallel processing of unit tests available in
testthat edition 3 (which is opt-in)?
- Can I add `Config/testthat/parallel: true` to the DESCRIPTION without
breaking things?
    - ADDED the config to DESCRIPTION, and no longer need the reporter.
4. `.pre-commit-config.yaml` The defaults here are styling and
re-documenting, right? But we have workflows for those as well. Do we
need both? What is the role of all the dependencies listed in the file?
- I can leave this and not install the pre-commit hooks. Or i can delete
this
6. `.gitlab-ci.yml`, do I need this one since this is a GH repo? KEEP
THIS BECAUSE WE WILL NEED IT FOR VALIDATION.
7. Where are staged dependencies used? I see them in the R CMD Checks
yaml. Any other spaces? I don't think I need that infrastructure right
now: if I delete the file now, will it break existing workflows?
   - I CAN DELETE OR NOT DELETE
9. `build-check-install.yaml` 
- Can we discuss the details of `junit-xml` reports, so I can understand
what they are doing?
- What versions of R are checked? A project like cards need to be
checked on the last 4 versions of R.
- This checks one version of R only for now. This may change in the
future. For now, I'll probably just add the Posit checks for older
versions.
- The reports that are uploaded to the pkgdown site, are the just static
html files? If I prefer not to add a dropdown menu that persists all
every page of the website, can I link to them another way?
- I can delete these menus from the _pkgdown.yaml, and then add a link
via a fa-icon
- Is this built similarly to the admiral workflows based off images that
update every so often? If so, how frequently do they update? What if I
need to use a new feature from a package?
 10. I ❤️ `branch-cleanup`. Did I specify this option correctly?
 

![image](https://github.com/insightsengineering/cardx/assets/26774684/d4ee72ce-71c3-4526-9234-01d9ccba040e)

11. Does either the lintr or styler actually change my scripts or just
report when they are not properly linted/styled? If so, can that be
changed to just report?