Adding `ard_stats_anova()` (#107)
**What changes are proposed in this pull request?**
- `ard_stats_anova()` for calculating ANOVA results using
`stats::anova()`. (#12)

There are two implementations of this function via S3 methods.
1. The first is when a user calculates the `anova()` object and passes
it to `ard_stats_anova()`.
2. The second is when a user passes a data frame. We use that data frame
and information from the other arguments to construct the models and
pass the models to `anova()`

**Reference GitHub issue associated with pull request.** _e.g., 'closes
#<issue number>'_
closes #12


--------------------------------------------------------------------------------

Pre-review Checklist (if item does not apply, mark is as complete)
- [ ] **All** GitHub Action workflows pass with a :white_check_mark:
- [ ] PR branch has pulled the most recent updates from master branch:
`usethis::pr_merge_main()`
- [ ] If a bug was fixed, a unit test was added.
- [ ] If a new `ard_*()` function was added, it passes the ARD
structural checks from `cards::check_ard_structure()`.
- [ ] Code coverage is suitable for any new functions/features
(generally, 100% coverage for new code): `devtools::test_coverage()`
- [ ] Request a reviewer

Reviewer Checklist (if item does not apply, mark is as complete)

- [x] If a bug was fixed, a unit test was added.
- [x] Run `pkgdown::build_site()`. Check the R console for errors, and
review the rendered website.
- [x] Code coverage is suitable for any new functions/features:
`devtools::test_coverage()`

When the branch is ready to be merged:
- [ ] Update `NEWS.md` with the changes from this pull request under the
heading "`# cards (development version)`". If there is an issue
associated with the pull request, reference it in parentheses at the end
update (see `NEWS.md` for examples).
- [ ] **All** GitHub Action workflows pass with a :white_check_mark:
- [ ] Approve Pull Request
- [ ] Merge the PR. Please use "Squash and merge" or "Rebase and merge".

---------

Signed-off-by: Daniel Sjoberg <danield.sjoberg@gmail.com>
Co-authored-by: Emily de la Rua <59304861+edelarua@users.noreply.github.com>